azure_api_container/requirements.txt
azure-core==1.13.0
azure-storage-file-datalake==12.3.1
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
requests==2.22.0
urllib3==1.25.7
PyYAML
Flask==1.1.2